Hibernate Database WIP5

+++ src/main/java/pl/coderstrust/database/HibernateDatabase.java
@@ -1,6 +1,7 @@
 package pl.coderstrust.database;
 
 import org.springframework.dao.NonTransientDataAccessException;
+import pl.coderstrust.database.hibernate.HibernateInvoice;
 import pl.coderstrust.database.hibernate.InvoiceRepository;
 import pl.coderstrust.model.Invoice;
 
@@ -10,9 +11,11 @@ import java.util.Optional;
 
 public class HibernateDatabase implements Database {
     private final InvoiceRepository invoiceRepository;
+    private final HibernateModelMapper modelMapper;
 
-    public HibernateDatabase(InvoiceRepository invoiceRepository) {
+    public HibernateDatabase(InvoiceRepository invoiceRepository, HibernateModelMapper modelMapper) {
         this.invoiceRepository = invoiceRepository;
+        this.modelMapper = modelMapper;
     }
 
     //stwórz new HibernateInvoice
@@ -26,7 +29,8 @@ public class HibernateDatabase implements Database {
         if (invoice == null) {
             throw new IllegalArgumentException("Invoice cannot be null.");
         }
-        return invoiceRepository.save(invoice);
+        HibernateInvoice savedInvoice = invoiceRepository.save(modelMapper.mapInvoice(invoice));
+        return modelMapper.mapInvoice(savedInvoice);
     }
 
     @Override
@@ -34,21 +38,19 @@ public class HibernateDatabase implements Database {
         if (id == null) {
             throw new IllegalArgumentException("Invoice id cannot be null.");
         }
-//        if (!invoiceRepository.existsById(id)) {
-//            throw new DatabaseOperationException(String.format("There is no invoice with id: %s", id));
-//        }
+        if (!invoiceRepository.existsById(id)) {
+            throw new DatabaseOperationException(String.format("There is no invoice with id: %s", id));
+        }
         invoiceRepository.deleteById(id);
     }
 
     @Override
-    public Optional<Invoice> getById(Long id) throws DatabaseOperationException {
+    public Optional<Invoice> getById(Long id) {
         if (id == null) {
             throw new IllegalArgumentException("Id cannot be null.");
         }
-//        if (!invoiceRepository.existsById(id)) {
-//            throw new DatabaseOperationException(String.format("There is no invoice with id: %s", id));
-//        }
-        return Optional.ofNullable(invoiceRepository.getOne(id));
+        return null;
+        // return invoiceRepository.findById(id);
     }
 
     @Override
@@ -56,21 +58,22 @@ public class HibernateDatabase implements Database {
         if (number == null) {
             throw new IllegalArgumentException("Invoice number cannot be null");
         }
-        return invoiceRepository.findAll()
+        return null;
+       /* return invoiceRepository.findAll()
                 .stream()
                 .filter(invoice -> invoice.getNumber().equals(number))
-                .findFirst();
+                .findFirst();*/
     }
 
     @Override
     public Collection<Invoice> getAll() throws DatabaseOperationException {
         List<Invoice> invoiceList;
         try {
-            invoiceList = invoiceRepository.findAll();
+            return null;
+            /*  invoiceList = invoiceRepository.findAll();*/
         } catch (NonTransientDataAccessException e) {
             throw new DatabaseOperationException("An error occurred during deleting all invoices.", e);
         }
-        return invoiceList;
     }
 
     @Override
